fix : Dto, Entity에 Builder 클래스 수정해 적용, 불필요한 코드 삭제, domain
