--- src/main/java/com/devsong/server/domain/Category.java
@@ -1,5 +1,0 @@
-package com.devsong.server.domain;
-
-public enum Category {
-    PROJECT, STUDY, EXTRACURRICULARACTIVITY, INFORMATION, FREETALK
-}
