+++ src/main/java/com/devsong/server/post/entity/Category.java
@@ -1,0 +1,4 @@
+package com.devsong.server.post.entity;
+
+public enum Category {
+    PROJECT, STUDY, EXTRACURRICULARACTIVITY, INFORMATION, FREETALK}
