+++ src/main/java/com/devsong/server/post/entity/PostLike.java
@@ -1,30 +1,27 @@
-package com.devsong.server.domain;
+package com.devsong.server.post.entity;
 
+import com.devsong.server.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
 @Getter
+@Builder
+@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_like_id")
-    private Long id;
+    private Long id; //pk
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+    private User user; //유저
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
+    private Post post; //게시글
 
-    @Builder
-    public PostLike(User user, Post post) {
-        this.user = user;
-        this.post = post;
-    }
 }
